backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
swoop-flights
